change in adjacence matrix. from double to uint32_t

+++ sequential/include/Multigraph.hpp
@@ -17,7 +17,7 @@ typedef struct _Edge_t
 {
     uint32_t from;
     uint32_t to;
-    double weight;
+    uint32_t weight;
 
 } Edge_t;
 
@@ -26,8 +26,8 @@ class Multigraph
 private:
     vector<Vertex_t> vertices;
     vector<vector<uint32_t>> neighbors;
-    vector<vector<double>> adjacencyMatrix;
-    const static double INFINITY;
+    vector<vector<uint32_t>> adjacencyMatrix;
+    const static uint32_t INFINITY;
     
 public:
     Multigraph(/* args */);
@@ -38,11 +38,11 @@ public:
     void addEdge(uint32_t from, uint32_t to);
     void addVertices(vector<Vertex_t> &newVertices);
     void addEdges(vector<Edge_t> &edges);
-    vector<vector<double>> *getAdjacencyMatrix();
+    vector<vector<uint32_t>> *getAdjacencyMatrix();
     vector<vector<uint32_t>> *getNeighbors();
     bool isEulerian(vector<uint32_t> &oddVertices);
-    bool dijkstra(uint32_t start, uint32_t end, list<uint32_t> &outputPath, double &totalDistance);
-    bool hierholzer(vector<uint32_t> &outputPath, double &distance);
+    bool dijkstra(uint32_t start, uint32_t end, list<uint32_t> &outputPath, uint64_t &totalDistance);
+    bool hierholzer(vector<uint32_t> &outputPath, uint64_t &distance);
     void print();
 };
